+++ src/app/(tabs)/_layout.tsx
@@ -1,12 +1,11 @@
 import React from "react";
-import FontAwesome from "@expo/vector-icons/FontAwesome";
-import { Link, Tabs } from "expo-router";
-import { Pressable, SafeAreaView } from "react-native";
+import { Tabs } from "expo-router";
 
-import Colors from "@/src/constants/Colors";
+import { useClientOnlyValue } from "@/src/hooks/useClientOnlyValue";
+
+import FontAwesome from "@expo/vector-icons/FontAwesome";
 import { COLORS } from "@/src/constants/Colors";
 import { useColorScheme } from "@/src/hooks/useColorScheme";
-import { useClientOnlyValue } from "@/src/hooks/useClientOnlyValue";
 
 // You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
 function TabBarIcon(props: {
@@ -36,20 +35,20 @@ export default function TabLayout() {
           tabBarIcon: ({ color }) => (
             <TabBarIcon name="binoculars" color={color} />
           ),
-          headerRight: () => (
-            <Link href="/modal" asChild>
-              <Pressable>
-                {({ pressed }) => (
-                  <FontAwesome
-                    name="info-circle"
-                    size={25}
-                    color={Colors[colorScheme ?? "light"].text}
-                    style={{ marginRight: 15, opacity: pressed ? 0.5 : 1 }}
-                  />
-                )}
-              </Pressable>
-            </Link>
-          ),
+          // headerRight: () => (
+          //   <Link href="/modal" asChild>
+          //     <Pressable>
+          //       {({ pressed }) => (
+          //         <FontAwesome
+          //           name="info-circle"
+          //           size={25}
+          //           color={Colors[colorScheme ?? "light"].text}
+          //           style={{ marginRight: 15, opacity: pressed ? 0.5 : 1 }}
+          //         />
+          //       )}
+          //     </Pressable>
+          //   </Link>
+          // ),
         }}
       />
       <Tabs.Screen
